pyproject.toml
[project]
name = "group7-python-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexandr-1973",email = "144156771+Alexandr-1973@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-phonenumber-field[phonenumbers] (>=8.1.0,<9.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "supabase (>=2.15.3,<3.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "rich (>=14.0.0,<15.0.0)"
]

package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
